Use Minecraft player join event type
