+++ lib/index.js
@@ -10908,14 +10908,6 @@ import {
   world as world2
 } from "@minecraft/server";
 
-// src/router/events/classes/AddonActivateAfterEvent.ts
-var AddonActivateAfterEvent = class {
-};
-
-// src/router/events/classes/AddonDeactivateBeforeEvent.ts
-var AddonDeactivateBeforeEvent = class {
-};
-
 // src/utils/SeedRandom.ts
 var import_seedrandom = __toESM(require_seedrandom2(), 1);
 var SeedRandom = class {
@@ -11021,6 +11013,14 @@ var KairoRuntime = class {
     runTimeout: (cb, tick) => system.runTimeout(cb, tick),
     clearRun: (id) => system.clearRun(id)
   };
+};
+
+// src/router/events/classes/AddonActivateAfterEvent.ts
+var AddonActivateAfterEvent = class {
+};
+
+// src/router/events/classes/AddonDeactivateBeforeEvent.ts
+var AddonDeactivateBeforeEvent = class {
 };
 
 // src/router/errors/KairoRouterError.ts
