+++ src/router/events/EventRegistry.ts
@@ -1,6 +1,6 @@
 import { KairoRouterError, KairoRouterErrorReason } from "../errors/KairoRouterError";
 import { InternalEvent } from "./types/InternalEvent";
-import type { KairoEventMap } from "./types/KairoEventMap";
+import type { KairoEventMap } from "../../minecraft/KairoEventMap";
 
 export class EventRegistry<E extends KairoEventMap> {
     private afterStore = new Map<keyof E["after"], InternalEvent<any>>();
