--- src/router/events/types/KairoEventMap.ts
@@ -1,13 +1,0 @@
-import { AddonActivateAfterEvent } from "../classes/AddonActivateAfterEvent";
-import { AddonDeactivateBeforeEvent } from "../classes/AddonDeactivateBeforeEvent";
-import type { KairoPlayerJoinAfterEvent } from "../classes/KairoPlayerJoinAfterEvent";
-
-export interface KairoEventMap {
-    readonly after: {
-        readonly addonActivate: AddonActivateAfterEvent;
-        readonly playerJoin: KairoPlayerJoinAfterEvent;
-    };
-    readonly before: {
-        readonly addonDeactivate: AddonDeactivateBeforeEvent;
-    };
-}
